+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Identity column type
 - Tsvector column type
 - Functional Indexes
-- Column Index Options 
+- Column Index Options
+- Column compression modifier
 
 ## [0.4.0] - 2021-09-18
 ### Added